Rename theme to components
Components makes more sense as a name.
