+++ src/components/styles.tsx
